Avoid generating `UserId`s in the client API.
 It is the responsibility of the server side code to manage user accounts and therefore their IDs.
 As of now we need `UserId` generation only in tests.
 We cannot prohibit creation of `UserId` instances in the client API. But we should not encourage doing it either. So, I moved the factory method for `UserId` generation into the `Tests` utility class.
