+++ client/src/test/java/org/spine3/protobuf/TypeUrlShould.java
@@ -30,7 +30,7 @@ import org.junit.Test;
 import org.spine3.base.Command;
 import org.spine3.base.CommandValidationError;
 import org.spine3.client.CommandFactory;
-import org.spine3.client.test.TestCommandFactory;
+import org.spine3.test.TestCommandFactory;
 import org.spine3.test.Tests;
 import org.spine3.users.UserId;
 import org.spine3.validate.internal.RequiredOption;
