+++ client/src/test/java/org/spine3/testdata/TestCommandContextFactory.java
@@ -30,8 +30,8 @@ import org.spine3.users.TenantId;
 import org.spine3.users.UserId;
 
 import static org.spine3.base.Identifiers.newUuid;
-import static org.spine3.client.UserUtil.newUserId;
 import static org.spine3.protobuf.Timestamps.getCurrentTime;
+import static org.spine3.test.Tests.newUserId;
 import static org.spine3.time.ZoneOffsets.UTC;
