+++ server/src/test/java/org/spine3/server/command/CommandBusShould.java
@@ -40,7 +40,6 @@ import org.spine3.base.FailureThrowable;
 import org.spine3.base.Response;
 import org.spine3.base.Responses;
 import org.spine3.client.CommandFactory;
-import org.spine3.client.test.TestCommandFactory;
 import org.spine3.protobuf.Durations;
 import org.spine3.server.command.error.CommandException;
 import org.spine3.server.command.error.InvalidCommandException;
@@ -49,6 +48,7 @@ import org.spine3.server.event.EventBus;
 import org.spine3.server.storage.memory.InMemoryStorageFactory;
 import org.spine3.server.type.CommandClass;
 import org.spine3.server.users.CurrentTenant;
+import org.spine3.test.TestCommandFactory;
 import org.spine3.test.Tests;
 import org.spine3.test.command.AddTask;
 import org.spine3.test.command.CreateProject;
