+++ server/src/test/java/org/spine3/server/command/Given.java
@@ -32,8 +32,8 @@ import org.spine3.test.command.StartProject;
 import org.spine3.users.UserId;
 
 import static org.spine3.base.Identifiers.newUuid;
-import static org.spine3.client.UserUtil.newUserId;
 import static org.spine3.protobuf.Timestamps.getCurrentTime;
+import static org.spine3.test.Tests.newUserId;
 import static org.spine3.testdata.TestCommandContextFactory.createCommandContext;
 
 /* package */ class Given {
