+++ server/src/test/java/org/spine3/testdata/TestEventContextFactory.java
@@ -33,9 +33,9 @@ import org.spine3.server.integration.IntegrationEventContext;
 import org.spine3.users.UserId;
 
 import static org.spine3.base.Identifiers.newUuid;
-import static org.spine3.client.UserUtil.newUserId;
 import static org.spine3.protobuf.Timestamps.getCurrentTime;
 import static org.spine3.protobuf.Values.newStringValue;
+import static org.spine3.test.Tests.newUserId;
 
 
 /**
